scripts-checking security related with event 4625
In this folder and this 10 scripts I combined some techniques learned from PREVIOUS powershell scripts that helped me to develop THESE SCRIPTS AND CAN HELP TO DEVELOP new scripts for other security events. In this 10 scripts ARE VERY USEFUL COMMANDS AND BLOCKS OF POWERSHELL INSTRUCTIONS TO ANALYZE SOME EVENTS IN SECURITY LOG FOR WINDOWS SERVER 2012 R2.